created register post request & refactor some issues

+++ src/main/java/com/example/demoJDBC/Employee.java
@@ -4,6 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
@@ -12,5 +15,5 @@ public class Employee {
      private int id ;
      private String name;
      private String position;
-     private double salary;
+     private BigDecimal salary;
 }
